fix: Correct SQLite sum error and update UI styles
This commit addresses user feedback on the initial redesign:

- fix(HomePage): Resolves the "SQLite cannot apply aggregate operator 'Sum' on expressions of type 'decimal'" error by casting the value to a `double` within the LINQ query.
- style(Styles): Updates the global `Button` style in `Styles.xaml` to use a white foreground color for better contrast and readability against the green accent backgrounds.
- style(HomePage): Removes local overrides on `HomePage.xaml` buttons to ensure they correctly inherit the updated global style.
- style(AddTransactionWindow): Completely refactors `AddTransactionWindow.xaml` to use the global `ResourceDictionary`, ensuring UI consistency with the new "spring forest" theme. Removes hardcoded colors and local styles.

+++ src/Kopilka.FinanceManager/Views/Pages/HomePage.xaml.cs
@@ -52,9 +52,9 @@ namespace Kopilka.FinanceManager.Views.Pages
 
         private async System.Threading.Tasks.Task UpdateTotalBalanceAsync(KopilkaDbContext dbContext)
         {
-            var totalBalance = await dbContext.Accounts
+            var totalBalance = (decimal)await dbContext.Accounts
                                               .Where(a => a.UserId == _currentUser.Id)
-                                              .SumAsync(a => a.Balance);
+                                              .SumAsync(a => (double)a.Balance);
             TotalBalanceTextBlock.Text = $"{totalBalance:N2} ₽";
         }
